b_past_life_mcp/requirements.txt
pandas>=2.0.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

